# target/classes/openapi/user-api.yaml
openapi: 3.0.3
info:
  title: User Management API
  version: 1.0.0
  x-java-interface-name: UserApi
paths:
  /api/users/register:
    post:
      operationId: register
      summary: 用户注册
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegisterRequest'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/users/login:
    post:
      operationId: login
      summary: 用户登录（支持用户名/邮箱/手机号）
      tags:
        - User
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

  /api/users/password:
    put:
      operationId: updatePassword
      summary: 修改密码（需登录）
      tags:
        - User
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePasswordRequest'
      responses:
        '200':
          description: 密码修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/users/{userId}:
    delete:
      operationId: deleteUser
      summary: 删除用户（管理员可删除所有用户，普通用户仅可删除自己）
      tags:
        - User
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 删除成功

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserRegisterRequest:
      type: object
      required:
        - username
        - name
        - email
        - password
      properties:
        username:
          type: string
          minLength: 4
          maxLength: 50
        name:
          type: string
          maxLength: 100
        email:
          type: string
          format: email
        phone:
          type: string
          pattern: '^1[3-9]\d{9}$' # 手机号正则（可选）
        password:
          type: string
          minLength: 8
          format: password
        gender:
          type: string
          enum: [M, F, U]
          default: U
        birthday:
          type: string
          format: date
        occupation:
          type: string
        signature:
          type: string

    UserLoginRequest:
      type: object
      required:
        - account
        - password
      properties:
        account:
          type: string # 用户名/邮箱/手机号
        password:
          type: string
          format: password

    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT令牌
        userId:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string # 角色（如user、admin等）

    UpdatePasswordRequest:
      type: object
      required:
        - oldPassword
        - newPassword
      properties:
        oldPassword:
          type: string
          format: password
        newPassword:
          type: string
          minLength: 8
          format: password

    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          description: 状态码（200成功，其他失败）
        message:
          type: string
          description: 提示信息
        data:
          type: object